requirements.txt
google-genai
python-dotenv
networkx
